+++ web-frame/src/app-shared/config/config.module.ts
@@ -10,30 +10,11 @@ import {
   Optional,
   SkipSelf,
 } from '@angular/core';
-import { ExtraOptions, RouterModule } from '@angular/router';
 
 import { ENVIRONMENT } from './defaults/environment';
 import { IEnvironmentConfig, validateEnvironmentConfig } from './interfaces/environment-config.interface';
 
-// NOTE: The initialNavigation is delayed until after AppFeaturesInitializerService.
-const ROUTES = [];
-
-const ROUTER_CONFIG: ExtraOptions = {
-  initialNavigation: 'disabled',
-  onSameUrlNavigation: 'reload',
-  paramsInheritanceStrategy: 'always',
-  relativeLinkResolution: 'corrected',
-  urlUpdateStrategy: 'eager',
-  useHash: true,
-};
-
 @NgModule({
-  exports: [
-    RouterModule,
-  ],
-  imports: [
-    RouterModule.forRoot(ROUTES, ROUTER_CONFIG),
-  ],
 })
 export class ConfigModule {
   constructor(
